allow reuse of Image, ImageLayer and ImageManifest types
Issue: AAH-2522
